requirements.txt
datasets>=2.18.0
pandas>=2.0.0
pytest>=8.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.28.0
groq>=0.4.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

